+++ lib/scrum.rb
@@ -4,8 +4,10 @@ end
 require_relative 'scrum/card_type_detection'
 require_relative 'scrum/sprint_board'
 require_relative 'scrum/sprint_planning_board'
+require_relative 'scrum/scrum_boards'
 
 require_relative 'scrum/priority_name'
 require_relative 'scrum/prioritizer'
 require_relative 'scrum/sprint_cleaner'
 require_relative 'scrum/backlog_mover'
+require_relative 'scrum/creator'